backend/requirements.txt
gradio
langchain 
openai
python-dotenv
langchain_community
fastapi
langchain_openai
httpx